test: export utils for setting the MockAgent and assertions for NodeSecure standard format
